fix: switch to long data type for upload progress tracking

+++ Blackbox.Client/Events/UploadProgressEventArgs.cs
@@ -4,14 +4,14 @@ namespace Blackbox.Client.Events
 {
     public class UploadProgressEventArgs : EventArgs
     {
-        public UploadProgressEventArgs(int bytesTotal, int bytesProcessed)
+        public UploadProgressEventArgs(long bytesTotal, long bytesProcessed)
         {
             BytesTotal = bytesTotal;
             BytesProcessed = bytesProcessed;
         }
 
-        public int BytesTotal { get; private set; }
-        public int BytesProcessed { get; private set; }
+        public long BytesTotal { get; private set; }
+        public long BytesProcessed { get; private set; }
         public int Percent => (int)Math.Round(((double) BytesProcessed / BytesTotal) * 100);
     }
 }
